+++ tests/test_job_compiler_blender_render_progressive.py
@@ -5,6 +5,8 @@ from unittest import mock
 
 from bson import ObjectId, tz_util
 
+from pillar.tests import common_test_data as ctd
+from abstract_flamenco_test import AbstractFlamencoTest
 from test_job_compilers import JobDocForTesting
 
 
@@ -28,7 +30,7 @@ class BlenderRenderProgressiveTest(unittest.TestCase):
         self.assertEqual(expected, list(cg), 're-iterating "cg" should work')
 
         cg = brp.ChunkGenerator(400, 100, uncapped_chunks=4)
-        samples = [(end-start+1) for start, end in cg]
+        samples = [(end - start + 1) for start, end in cg]
         self.assertEqual([10, 62, 82, 82, 82, 82], samples)
 
         cg = brp.ChunkGenerator(30, 3000, uncapped_chunks=3)
@@ -484,3 +486,81 @@ class BlenderRenderProgressiveTest(unittest.TestCase):
         task_manager.api_set_task_status_for_job.assert_called_with(
             job_doc['_id'], 'under-construction', 'queued', now=mock_now)
         job_manager.api_set_job_status(job_doc['_id'], 'under-construction', 'queued', now=mock_now)
+
+
+class CreateDeferredTest(AbstractFlamencoTest):
+    def setUp(self):
+        super().setUp()
+
+        self.create_standard_groups()
+
+        # Create a timestamp before we start mocking datetime.datetime.
+        self.created = datetime.datetime(2018, 7, 6, 11, 52, 33, tzinfo=tz_util.utc)
+
+        mngr_doc, account, token = self.create_manager_service_account()
+        self.mngr_id = mngr_doc['_id']
+
+        self.create_user(user_id=24 * 'e',
+                         roles={'subscriber'},
+                         groups=[mngr_doc['owner'], ctd.EXAMPLE_ADMIN_GROUP_ID],
+                         token='owner-token')
+        with self.app.app_context():
+            self.mmngr.api_assign_to_project(self.mngr_id, self.proj_id, 'assign')
+
+    def test_create_deferred(self):
+        job_doc = {
+            '_id': ObjectId(24 * 'f'),
+            'settings': {
+                'frames': '1-5',
+                'chunk_size': 3,
+                'render_output': '/render/out/frames-######',
+                'fps': 5.3,
+                'format': 'EXR',
+                'blender_cmd': '/path/to/blender --enable-new-depsgraph',
+                'cycles_sample_count': 30,
+
+                # Effectively uncapped so that the number of tasks stays small.
+                # The actual capping is tested in test_chunk_generator() anyway.
+                'cycles_sample_cap': 30,
+            },
+            'job_type': 'blender-render-progressive',
+            'manager': str(self.mngr_id),
+            'name': 'Zzžžž',
+            'project': str(self.proj_id),
+            'start_paused': True,
+            'status': 'waiting-for-files',
+            'user': 24 * 'e',
+        }
+        resp = self.post('/api/flamenco/jobs', json=job_doc, auth_token='owner-token',
+                         expected_status=201)
+        job_id = ObjectId(resp.json['_id'])
+
+        job_url = f'/api/flamenco/jobs/{job_id}'
+        json_job = self.get(job_url, auth_token='owner-token').json
+        self.assertEqual('waiting-for-files', json_job['status'])
+
+        # Check that there are no tasks created.
+        with self.app.app_context():
+            tasks_coll = self.flamenco.db('tasks')
+            self.assertEqual([], list(tasks_coll.find({'job': job_id})))
+
+        # PATCH the job to get the task creation going.
+        filepath = '{shaman}/jo/b-id/scenes/someshot/somefile.blend'
+        self.patch(job_url,
+                   json={'op': 'construct', 'settings': {'filepath': filepath}},
+                   auth_token='owner-token',
+                   expected_status=204)
+
+        json_job = self.get(job_url, auth_token='owner-token').json
+        self.assertEqual('paused', json_job['status'])
+        self.assertEqual(filepath, json_job['settings']['filepath'])
+
+        with self.app.app_context():
+            tasks = list(tasks_coll.find({'job': job_id}))
+
+        self.assertEqual('{shaman}/jo/b-id/scenes/someshot/somefile.blend',
+                         tasks[1]['commands'][0]['settings']['filepath'])
+        for task in tasks:
+            self.assertEqual(job_id, task['job'])
+            self.assertEqual('paused', task['status'])
+            self.assertEqual(self.mngr_id, task['manager'])